dom/workers/test/test_bug1062920.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for navigator property override</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  function checkValues() {
    var worker = new Worker("bug1062920_worker.js");

    worker.onmessage = function(event) {
      is(event.data.appCodeName, navigator.appCodeName, "appCodeName should match");
      is(event.data.appName, navigator.appName, "appName should match");
      is(event.data.appVersion, navigator.appVersion, "appVersion should match");
      is(event.data.platform, navigator.platform, "platform should match");
      is(event.data.userAgent, navigator.userAgent, "userAgent should match");
      is(event.data.product, navigator.product, "product should match");
      runTests();
    };
  }

  function replaceAndCheckValues() {
    SpecialPowers.pushPrefEnv({"set": [
      ["general.appname.override", "appName overridden"],
      ["general.appversion.override", "appVersion overridden"],
      ["general.platform.override", "platform overridden"]
    ]}, checkValues);
  }

  var tests = [
    checkValues,
    replaceAndCheckValues
  ];

  function runTests() {
    if (tests.length == 0) {
      SimpleTest.finish();
      return;
    }

    var test = tests.shift();
    test();
  }

  SimpleTest.waitForExplicitFinish();
  runTests();

</script>
</pre>
</body>
</html>
